Added repositories and dbcontext with interceptor and initializer

+++ RPSSL.GameService/RPSSL.GameService.Domain/Models/LocalizationLabel.cs
@@ -3,7 +3,7 @@
 public class LocalizationLabel
 {
 	public required string Key { get; set; }
-	public required int LanguageId { get; set; }
+	public required short LanguageId { get; set; }
 	public required string Value { get; set; }
 
 	public virtual Language Language { get; set; } = null!;
